feat(exam): implement snooze functionality for alarms
Modified `RingingAlarmActivity.kt` to handle the snooze button click, sending a broadcast to `AlarmReceiver` to snooze the alarm.
Updated `ExamViewModel.kt` and `ExamScreen.kt` to dynamically use exam type for loading subtypes.
Adjusted `AlarmForegroundService.kt` to pass the correct `requestCode` and `label` for snooze and cancel intents.
Enhanced `AlarmReceiver.kt` to handle the "alarm.ACTION_SNOOZE" action by stopping the foreground service and updating the alarm time in the database.
Made the `time` property in `AlarmEntity.kt` a `var` to allow modification for snoozing.